iterations/1-mvp/requirements.txt
haystack-ai
accelerate
"sentence-transformers>=3.0.0"
"datasets>=2.6.1"
qdrant-haystack
markdown-it-py
mdit_plain
pypdf